refactor(struct): move reconstruction adapter to app research
